refactor(dfpt): extract DFPT_PW q=0 solves (pos + efield SCF) into dfpt_pw_q0.cpp
Moves DFPT_PW::Impl::solve_pos_resp (complexity 31) and
solve_efield_resp (complexity 25) from dfpt_pw.cpp into the new
dfpt_pw_q0.cpp translation unit, assisted by five private Impl
helpers: vel_diag_part, vel_nl_per_atom, pos_per_band_solve,
efield_per_band_solve, stash_dpsi_efield. The assemble_v_sc helper
now lives in dfpt_pw_solve.cpp and is reused in both the
displacement and efield SCF fixed points instead of duplicating
the Hartree+XC kernel assembly across two drivers.

After this commit, dfpt_pw.cpp drops from 1098 LOC to 156 LOC and
contains only the thin public pimpl wrapper, DFPT_PW::Impl::wired(),
and the small accessor / setter group; the core implementation is
spread across four single-responsibility translation units
(dfpt_pw_{init,run,solve,q0}.cpp), each under 300 LOC. All public
API behavior, INPUT parameters, and test references are 100%
unchanged.

Verification commands:
  cmake --build build -j 30                         -> PASS
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_pw_run_test                -> PASS
  cmake --build build_max_para_test -j 30 \
    --target MODULE_DFPT_{pert_serial,phon_serial,\
    q0_serial,rho_serial}                           -> PASS all
  OMP_NUM_THREADS=1 ctest --test-dir build_max_para_test \
    -R MODULE_DFPT                                  -> 8/8 Passed (5.13 s)
  python3 tools/03_code_analysis/agent_governance_check.py \
    --staged                                        -> 0 deterministic
                                                       errors (2 PR-body
                                                       warnings only; same
                                                       API, same behavior)
