+++ source/source_pw/module_dfpt/CMakeLists.txt
@@ -5,6 +5,7 @@ set(SOURCES
     dfpt_pw_init.cpp
     dfpt_pw_run.cpp
     dfpt_pw_solve.cpp
+    dfpt_pw_q0.cpp
     dfpt_pw_data.cpp
     dfpt_kq_basis.cpp
     dfpt_pert.cpp
